update proto to account for moved configstream rpc in agent
