+++ lib/datadog-protos/proto/datadog/remoteagent/remoteagent.proto
@@ -3,9 +3,9 @@ syntax = "proto3";
 package datadog.remoteagent;
 
 
-import "google/protobuf/struct.proto";
 
 option go_package = "pkg/proto/pbgo/core"; // golang
+
 message StatusSection {
   map<string, string> fields = 1;
 }
@@ -87,27 +87,3 @@ message GetTelemetryResponse {
   }
 }
 
-message ConfigSetting {
-  string source = 1;
-  string key = 2;
-  google.protobuf.Value value = 3;
-}
-
-message ConfigSnapshot {
-  string origin = 1;
-  int32 sequence_id = 2;
-  repeated ConfigSetting settings = 3;
- }
-
-message ConfigUpdate {
-  string origin = 1;
-  int32 sequence_id = 2;
-  ConfigSetting setting = 3;
-}
-
-message ConfigEvent {
-  oneof event {
-    ConfigSnapshot snapshot = 1;
-    ConfigUpdate update = 2;
-  }
-}
